refactor: add global types folder, type files for webview, server, commands... import from index.types @ types
