+++ src/types/commands.types.d.ts
@@ -1,0 +1,16 @@
+import * as vscode from 'vscode';
+import * as WebviewTypes from './webview.types';
+
+export interface Compliance {
+  h1: WebviewTypes.AccessibilityNode | null;
+  skipLink: WebviewTypes.AccessibilityNode | null;
+}
+
+export interface A11yTreeCommands {
+  handleFetchTree: (
+    panel: vscode.WebviewPanel,
+    context: vscode.ExtensionContext,
+    url: string,
+    user: WebviewTypes.User
+  ) => Promise<void>;
+}
