+++ src/webview/components/URLInputForm.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { postMessage } from '../helpers/vscodeHelper';
 
-import { URLInputFormProps } from '../types';
+import { URLInputFormProps } from '../../types/index.types';
 
 function URLInputForm({ setPageResults, user }: URLInputFormProps) {
   const [url, setUrl] = useState<string>('http://127.0.0.1:5500/index.html');
