+++ src/webview/components/URLSelectionForm.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { postMessage } from '../helpers/vscodeHelper';
-import { URLSelectionFormProps } from '../types';
+import { URLSelectionFormProps } from '../../types/index.types';
 
 const URLSelectionForm = ({ setPageResults, user }: URLSelectionFormProps) => {
   const [url, setUrl] = useState<string>('http://127.0.0.1:5500/index.html');
